Just use the basic `int` type for digit values
Using the `int8` type adds complication to the code because of the
required conversions to and from `int` for many standard library
functions, and requires separate instantations of generic types and
functions.  The little bit of memory saved by using the shorter type
is unlikely to out-weigh the additional complexity, especially after
taking memory alignment rules into account.

+++ internal/puzzle/reader.go
@@ -36,7 +36,7 @@ func (p *Puzzle) processRow(row int, line string) {
 	for col := range 9 {
 		val := line[col] - 48
 		if val >= 1 && val <= 9 {
-			p.GivenValue(row, col, int8(val))
+			p.GivenValue(row, col, int(val))
 		}
 	}
 }
